Removes `internal` duplication of `GetContractSets`
