+++ deployment/keystone/changeset/types/forwarder.go
@@ -1,0 +1,6 @@
+package types
+
+const (
+	DeploymentBlockLabel = "deployment-block"
+	DeploymentHashLabel  = "deployment-hash"
+)
