+++ deployment/keystone/changeset/types/oracle_config.go
@@ -1,0 +1,72 @@
+package types
+
+import (
+	"encoding/json"
+	"fmt"
+	"time"
+)
+
+type OracleConfig struct {
+	MaxQueryLengthBytes       uint32
+	MaxObservationLengthBytes uint32
+	MaxReportLengthBytes      uint32
+	MaxOutcomeLengthBytes     uint32
+	MaxReportCount            uint32
+	MaxBatchSize              uint32
+	OutcomePruningThreshold   uint64
+	UniqueReports             bool
+	RequestTimeout            time.Duration
+
+	DeltaProgressMillis               uint32
+	DeltaResendMillis                 uint32
+	DeltaInitialMillis                uint32
+	DeltaRoundMillis                  uint32
+	DeltaGraceMillis                  uint32
+	DeltaCertifiedCommitRequestMillis uint32
+	DeltaStageMillis                  uint32
+	MaxRoundsPerEpoch                 uint64
+	TransmissionSchedule              []int
+
+	MaxDurationQueryMillis          uint32
+	MaxDurationObservationMillis    uint32
+	MaxDurationShouldAcceptMillis   uint32
+	MaxDurationShouldTransmitMillis uint32
+
+	MaxFaultyOracles int
+}
+
+func (oc *OracleConfig) UnmarshalJSON(data []byte) error {
+	type aliasT OracleConfig
+	temp := &struct {
+		RequestTimeout string `json:"RequestTimeout"`
+		*aliasT
+	}{
+		aliasT: (*aliasT)(oc),
+	}
+	if err := json.Unmarshal(data, temp); err != nil {
+		return fmt.Errorf("failed to unmarshal OracleConfig: %w", err)
+	}
+
+	if temp.RequestTimeout == "" {
+		oc.RequestTimeout = 0
+	} else {
+		requestTimeout, err := time.ParseDuration(temp.RequestTimeout)
+		if err != nil {
+			return fmt.Errorf("failed to parse RequestTimeout: %w", err)
+		}
+		oc.RequestTimeout = requestTimeout
+	}
+
+	return nil
+}
+
+func (oc OracleConfig) MarshalJSON() ([]byte, error) {
+	type aliasT OracleConfig
+	return json.Marshal(&struct {
+		RequestTimeout string `json:"RequestTimeout"`
+		*aliasT
+	}{
+		RequestTimeout: oc.RequestTimeout.String(),
+		aliasT:         (*aliasT)(&oc),
+	})
+}
